refactor(translation): move all VersionBlock related components to a new folder
