+++ static/js/VersionBlock/VersionBlock.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import Sefaria from './sefaria/sefaria';
-import Util from './sefaria/util';
-import $ from './sefaria/sefariaJquery';
+import Sefaria from '../sefaria/sefaria';
+import Util from '../sefaria/util';
+import $ from '../sefaria/sefariaJquery';
 import Component from 'react-class';
-import {LoadingMessage} from "./Misc";
+import {LoadingMessage} from "../Misc";
 import VersionBlockHeader from "./VersionBlockHeader";
 import VersionBlockSelectButton from "./VersionBlockSelectButton";
 import VersionInformation from "./VersionInformation";
