+++ static/js/VersionBlock/VersionMetadata.jsx
